+++ Assets/Scripts/Buildings/ArcherFortress.cs
@@ -10,7 +10,7 @@ public class ArcherFortress : AttackTower
     public override int BaseDamage => 5;
     public override Alliances Alliance => Alliances.Player;
     public override Alliances Enemies => Alliances.Illigons;
-    public override int BaseRange => RangeOptions.Medium;
+    public override float BaseRange => RangeOptions.Medium;
     public override int NumProjectiles => 5;
     public override float ProjectileStartPostionRandomness => .3f;
     public override VerticalRegion AttackRegion => VerticalRegion.GroundAndAir;
