+++ Assets/Scripts/Buildings/FireTower.cs
@@ -6,7 +6,7 @@ public class FireTower : AttackTower
 {
     public override float BaseCooldown => AttackSpeed.Medium;
     public override int BaseDamage => 20;
-    public override int BaseRange => RangeOptions.Medium;
+    public override float BaseRange => RangeOptions.Medium;
     public override BuildingType Type => BuildingType.FireTower;
     public override Alliances Alliance => Alliances.Player;
     public override Alliances Enemies => Alliances.Illigons;
