+++ Assets/Scripts/Buildings/FlamethrowerTower.cs
@@ -6,7 +6,7 @@ public class FlamethrowerTower : AttackTower
 {
     public override float BaseCooldown => AttackSpeed.VeryVeryVeryFast;
     public override int BaseDamage => 1;
-    public override int BaseRange => RangeOptions.Short;
+    public override float BaseRange => RangeOptions.Short;
     public override VerticalRegion AttackRegion => VerticalRegion.Ground;
     public override BuildingType Type => BuildingType.FlamethrowerTower;
     public override Alliances Alliance => Alliances.Player;
